requirements.txt
numpy>=1.26
PyYAML>=6.0
torch>=2.4
torchvision>=0.19
tqdm>=4.66
opencv-python-headless>=4.8
matplotlib>=3.8
Pillow>=10.0
tensorboard>=2.16
